+++ product/src/main/java/com/sparta/hotdeal/product/application/service/ProductService.java
@@ -15,8 +15,8 @@ import com.sparta.hotdeal.product.domain.entity.product.Product;
 import com.sparta.hotdeal.product.domain.entity.product.ProductStatusEnum;
 import com.sparta.hotdeal.product.domain.entity.product.SubFile;
 import com.sparta.hotdeal.product.domain.repository.product.ProductRepository;
-import com.sparta.hotdeal.product.domain.repository.product.ProductRepositoryCustomImpl;
 import com.sparta.hotdeal.product.infrastructure.dtos.ResGetCompanyByIdDto;
+import com.sparta.hotdeal.product.infrastructure.repository.ProductRepositoryCustomImpl;
 import java.math.BigDecimal;
 import java.math.RoundingMode;
 import java.util.List;
